removed pix key on bcb

+++ src/main/kotlin/br/com/zupedu/pix/RemovePixService.kt
@@ -1,6 +1,8 @@
 package br.com.zupedu.pix
 
 import br.com.zupedu.PixKeyRemoveRequest
+import br.com.zupedu.pix.externalConnections.bcb.ClientBcb
+import br.com.zupedu.pix.externalConnections.bcb.requests.DeletePixKeyRequest
 import br.com.zupedu.pix.externalConnections.itau.ClientItau
 import br.com.zupedu.pix.repository.PixRepository
 import br.com.zupedu.shared.handlingErrors.exceptions.KeyPixNotExistingException
@@ -17,7 +19,8 @@ import javax.validation.Valid
 @Singleton
 class RemovePixService(
     @Inject val pixRepository: PixRepository,
-    @Inject val clientItau: ClientItau
+    @Inject val clientItau: ClientItau,
+    @Inject val clientBcb: ClientBcb,
 ) {
 
     fun remove(@Valid request: PixKeyRemoveRequest) {
@@ -33,6 +36,15 @@ class RemovePixService(
         val pixKey = optionalPixKey.get()
         if(!pixKey.pixBelongsToTheClient(request.idClient))
             throw IllegalStateException("This pix key not belongs to this client")
+
+        val deletePixKeyRequest = DeletePixKeyRequest(pixKey.valueKey, pixKey.institution.ispb)
+        val responseKeyBcb = clientBcb.removeKey(pixKey.valueKey, deletePixKeyRequest)
+
+        if (responseKeyBcb.status == HttpStatus.NOT_FOUND)
+            throw IllegalStateException("This pix not found on system BCB")
+
+        if (responseKeyBcb.status == HttpStatus.FORBIDDEN)
+            throw IllegalStateException("Participant is not allowed to access this resource")
 
         pixRepository.deleteById(pixKey.id)
     }
